requirements.txt
kafka-python
pyspark
streamlit
psycopg2-binary
pandas